# docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://thinker.local:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/discovery/status": {
      "get": {
        "tags": [
          "device-discovery-controller"
        ],
        "operationId": "getDiscoveryStatus",
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorData"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorData"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ValidationErrorModel"
                      }
                    },
                    {
                      "$ref": "#/components/schemas/ErrorData"
                    },
                    {
                      "type": "object"
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorData"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoveryStatusModel"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "device-discovery-controller"
        ],
        "operationId": "setDiscoveryActive",
        "parameters": [
          {
            "name": "setActive",
            "in": "query",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorData"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorData"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ValidationErrorModel"
                      }
                    },
                    {
                      "$ref": "#/components/schemas/ErrorData"
                    },
                    {
                      "type": "object"
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorData"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoveryStatusModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/devices/{id}": {
      "get": {
        "tags": [
          "device-controller"
        ],
        "operationId": "getDevice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorData"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorData"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ValidationErrorModel"
                      }
                    },
                    {
                      "$ref": "#/components/schemas/ErrorData"
                    },
                    {
                      "type": "object"
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorData"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Device"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "device-controller"
        ],
        "operationId": "deleteDevice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorData"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorData"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ValidationErrorModel"
                      }
                    },
                    {
                      "$ref": "#/components/schemas/ErrorData"
                    },
                    {
                      "type": "object"
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorData"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          }
        }
      },
      "patch": {
        "tags": [
          "device-controller"
        ],
        "operationId": "updateDevice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorData"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorData"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ValidationErrorModel"
                      }
                    },
                    {
                      "$ref": "#/components/schemas/ErrorData"
                    },
                    {
                      "type": "object"
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorData"
                }
              }
            }
          },
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/discovery/connect/{address}": {
      "post": {
        "tags": [
          "device-discovery-controller"
        ],
        "operationId": "connectDevice",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorData"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorData"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ValidationErrorModel"
                      }
                    },
                    {
                      "$ref": "#/components/schemas/ErrorData"
                    },
                    {
                      "type": "object"
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorData"
                }
              }
            }
          },
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/discovery": {
      "get": {
        "tags": [
          "device-discovery-controller"
        ],
        "operationId": "getDevices",
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorData"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorData"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ValidationErrorModel"
                      }
                    },
                    {
                      "$ref": "#/components/schemas/ErrorData"
                    },
                    {
                      "type": "object"
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorData"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DiscoveredDevice"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/devices": {
      "get": {
        "tags": [
          "device-controller"
        ],
        "operationId": "getAllDevicesWithLatestReport",
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorData"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorData"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ValidationErrorModel"
                      }
                    },
                    {
                      "$ref": "#/components/schemas/ErrorData"
                    },
                    {
                      "type": "object"
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorData"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/x-ndjson": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeviceModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/devices/{id}/{actionName}": {
      "post": {
        "tags": [
          "device-controller"
        ],
        "operationId": "invokeDeviceAction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "actionName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorData"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorData"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ValidationErrorModel"
                      }
                    },
                    {
                      "$ref": "#/components/schemas/ErrorData"
                    },
                    {
                      "type": "object"
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorData"
                }
              }
            }
          },
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/discovery/restart": {
      "post": {
        "tags": [
          "device-discovery-controller"
        ],
        "operationId": "restartDiscovery",
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorData"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorData"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ValidationErrorModel"
                      }
                    },
                    {
                      "$ref": "#/components/schemas/ErrorData"
                    },
                    {
                      "type": "object"
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorData"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoveryStatusModel"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ErrorData": {
        "required": [
          "description",
          "descriptionKey",
          "httpStatus",
          "timestamp"
        ],
        "type": "object",
        "properties": {
          "httpStatus": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string"
          },
          "descriptionKey": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ValidationErrorModel": {
        "required": [
          "dataPath",
          "message"
        ],
        "type": "object",
        "properties": {
          "dataPath": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "DiscoveryStatusModel": {
        "required": [
          "isActive"
        ],
        "type": "object",
        "properties": {
          "isActive": {
            "type": "boolean"
          }
        }
      },
      "Device": {
        "required": [
          "actions",
          "address",
          "connectType",
          "description",
          "deviceClass",
          "name",
          "reportTypes",
          "status"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "WAITING_CONFIGURATION",
              "CONFIGURED",
              "FAILED_TO_CONFIGURE",
              "BLOCKED"
            ]
          },
          "connectType": {
            "type": "string",
            "enum": [
              "AUTO",
              "MANUAL"
            ]
          },
          "configurationHash": {
            "type": "integer",
            "format": "int32"
          },
          "actions": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DeviceAction"
            }
          },
          "deviceClass": {
            "type": "string"
          },
          "reportTypes": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "DeviceAction": {
        "required": [
          "description",
          "displayName",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "DeviceDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
      "DeviceReport": {
        "required": [
          "deviceId",
          "reportData",
          "reportType"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "deviceId": {
            "type": "string"
          },
          "reportType": {
            "type": "string"
          },
          "reportData": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "dateTimeCreated": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DiscoveredDevice": {
        "required": [
          "address",
          "discoveredAt",
          "knownDevice",
          "name",
          "rssi"
        ],
        "type": "object",
        "properties": {
          "address": {
            "type": "string"
          },
          "discoveredAt": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string"
          },
          "rssi": {
            "type": "integer",
            "format": "int32"
          },
          "knownDevice": {
            "type": "boolean"
          }
        }
      },
      "DeviceModel": {
        "required": [
          "actions",
          "address",
          "connectType",
          "description",
          "id",
          "name",
          "status"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "deviceClass": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "WAITING_CONFIGURATION",
              "CONFIGURED",
              "FAILED_TO_CONFIGURE",
              "BLOCKED"
            ]
          },
          "connectType": {
            "type": "string",
            "enum": [
              "AUTO",
              "MANUAL"
            ]
          },
          "actions": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DeviceAction"
            }
          },
          "latestReport": {
            "$ref": "#/components/schemas/DeviceReport"
          }
        }
      }
    }
  }
}